refactor(step2): Turn과 InitialTurn 분리
